Add bootstrapping guide for k8s cluster
This update introduces a new guide to help bootstrap Kubernetes cluster from scratch if required. It contains clear instructions on which commands to run to rebuild the cluster. Also, provides tips on how to handle secret automation in the event of credential loss. The final note mentions a potential future improvement with external-secrets.